// src/app/modules/common-regobs-api/model/compression-test-edit-model.ts
/**
 * RegObs API, build: 20251103.4 - commit: 59238037
 *
 * Contact: regobs@nve.no
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CompressionTestEditModel { 
    /**
     * The CompressionTestKDV unique identifier
     */
    CompressionTestTID?: number | null;
    /**
     * TapsFracture
     */
    TapsFracture?: number | null;
    /**
     * TapsFullPropagation
     */
    TapsFullPropagation?: number | null;
    /**
     * The PropagationKD unique identifier
     */
    PropagationTID?: number | null;
    /**
     * FractureDepth
     */
    FractureDepth?: number | null;
    /**
     * PST X distance (in meters)
     */
    PstX?: number | null;
    /**
     * PST Y distance (in meters)
     */
    PstY?: number | null;
    /**
     * The StabilityEvalKD unique identifier
     */
    StabilityEvalTID?: number | null;
    /**
     * The ComprTestFractureKD unique identifier
     */
    ComprTestFractureTID?: number | null;
    /**
     * Percentage of block that released in Rutchblock test
     */
    RbRelease?: number | null;
    /**
     * Comment
     */
    Comment?: string | null;
    IncludeInSnowProfile?: boolean | null;
}